simplifies code for creating displays and icons
icons are now created using their own functions so that i did
not have to re write the code to make a trash icon for each item or
project. for trash icons specifically they will have different behavior
depending on what calls them (ie project or item or verbose item)

additionally this commit begins to make the structure of projects and
items consisten across functions. this was necessary so that
the buttons could work regardless of where they were in the html
but also so that in a future commit classes can be re used for each
to simplify the css