will finalize the date window display
additionally, checked items will now always display as completed
even when changing displays.

the trashIcon event listeners have been simplified so there is one for
projects and one for items which now makes use of the remove() method so
it can be used in any display.

the choice to have any new item input trigger a regeneration of the
display to the tiled view was also made to simplify the code. in future
versions the ability to add directly to a verbose project or to the
today window will be added