implement cloud storage
cloud storage through the use of firebase and firestore is now
implemented

a collection for each user is created for any added items and
items are found through sorting through the user collection
and finding the specific item ID within that collection.

Randomized userID's and documenet references were chosen so that
in the event an itemID needed to be changed the document reference
could remain the same.